changing queries

+++ backend/documents/documents/src/main/java/io/writeopia/api/documents/documents/repository/DocumentSqlBeDao.kt
@@ -704,7 +704,7 @@ class DocumentSqlBeDao(
                             }
                         )
 
-                        innerContent.position!!.toDouble() to storyStep.copy(dbPosition = innerContent.position?.toDouble())
+                        innerContent.position!! to storyStep.copy(dbPosition = innerContent.position?.toDouble())
                     }
 
                     Document(
